Currently casting double values to float so that it works, may change later

+++ PNoiseGenerator.pde
@@ -2,7 +2,7 @@ public class PNoiseGenerator implements INoiseGenerator{
  
   public double perlin(int seed, double x, double y){
     noiseSeed(seed);
-    return noise(x, y);
+    return noise((float)x, (float)y);
   }
   
 }
